Removing unrelated code
