+++ package.json
@@ -1,91 +1,57 @@
 {
-  "name": "tinymce",
-  "version": "5.10.9",
-  "private": true,
+  "name": "tinymce-paste-from-word-plugin",
+  "version": "0.1.0",
+  "description": "TinyMCE 6.x Plugin to add support for pasting from Microsoft Word documents",
+  "author": "Pangaea Information Technologies, Ltd.",
+  "license": "LGPL-2.1",
+  "main": "dist/paste-from-word.js",
   "repository": {
     "type": "git",
-    "url": "https://github.com/tinymce/tinymce"
+    "url": "https://github.com/pangaeatech/tinymce-paste-from-word-plugin"
   },
-  "description": "TinyMCE rich text editor",
-  "author": "Tiny Technologies, Inc",
+  "keywords": [
+    "tinymce",
+    "Microsoft",
+    "Word",
+    "paste",
+    "plugin"
+  ],
   "bugs": {
-    "url": "https://github.com/tinymce/tinymce/issues"
+    "url": "https://github.com/pangaeatech/tinymce-paste-from-word-plugin/issues"
   },
-  "license": "LGPL-2.1",
   "engines": {
-    "node": ">=0.10.26"
+    "node": ">=16.0.0",
+    "npm": ">=8.0.0"
   },
-  "main": "tinymce.js",
   "scripts": {
+    "build": "ncc build src/main.ts",
+    "lint": "eslint . --ext .ts",
+    "test": "jest",
+    "coverage": "jest --coverage",
     "tsc": "tsc -b",
-    "watch": "tsc -b -w",
-    "test": "grunt test",
-    "lint": "grunt eslint",
-    "silver-test": "grunt bedrock-auto:silver",
-    "silver-test-manual": "grunt bedrock-manual:silver"
+    "watch": "tsc -b -w"
   },
   "dependencies": {
-    "dompurify": "2.3.8",
-    "tslib": "^2.0.0"
+    "@ephox/katamari": "^9.1.5",
+    "dompurify": "3.0.6",
+    "tinymce": "^6.8.1"
   },
   "devDependencies": {
-    "@ephox/bedrock-client": "^12.3.2",
-    "@ephox/bedrock-server": "^12.4.0",
-    "@ephox/oxide-icons-tools": "^2.2.2",
-    "@ephox/swag": "^4.4.0",
-    "@ephox/wrap-jsverify": "^2.0.1",
-    "@ephox/wrap-prismjs": "^1.29.1-patchfix.3558",
-    "@ephox/wrap-promise-polyfill": "^2.2.1",
-    "@tinymce/eslint-plugin": "^1.9.1",
-    "@tinymce/moxiedoc": "^0.1.0",
-    "@types/chai": "^4.2.15",
-    "awesome-typescript-loader": "^5.2.0",
-    "chai": "^4.3.3",
-    "chalk": "^4.1.0",
-    "emojilib": "^2.4.0",
-    "eslint": "^7.0.0",
-    "eslint-plugin-notice": "^0.9.10",
-    "fast-check": "^2.23.0",
-    "grunt": "^1.0.2",
-    "grunt-contrib-clean": "^2.0.0",
-    "grunt-contrib-concat": "^1.0.1",
-    "grunt-contrib-copy": "^1.0.0",
-    "grunt-contrib-less": "2.0.0",
-    "grunt-contrib-symlink": "^1.0.0",
-    "grunt-contrib-uglify": "^5.0.0",
-    "grunt-contrib-watch": "^1.1.0",
-    "grunt-eslint": "^23.0.0",
-    "grunt-nuget-pack": "^0.0.6",
-    "grunt-shell": "^3.0.1",
-    "grunt-webpack": "^4.0.2",
-    "gulp": "^4.0.0",
-    "gulp-clean": "^0.4.0",
-    "gulp-clean-css": "^4.0.0",
-    "gulp-connect": "^5.7.0",
-    "gulp-header": "^2.0.7",
-    "gulp-less": "^4.0.1",
-    "gulp-rename": "^2.0.0",
-    "gulp-sourcemaps": "^2.6.4",
-    "gulp-stylelint": "^13.0.0",
-    "lerna": "^3.18.4",
-    "less-plugin-autoprefix": "^2.0.0",
-    "load-grunt-tasks": "^5.1.0",
-    "moxie-zip": "0.0.4",
-    "npm-run-all": "^4.1.5",
-    "resolve": "^1.18.1",
-    "source-map-loader": "^1.1.0",
-    "string-replace-loader": "^2.1.1",
-    "stylelint": "^13.2.1",
-    "stylelint-config-standard": "^20.0.0",
-    "stylelint-order": "^4.0.0",
-    "ts-loader": "^6.2.2",
-    "tsconfig-paths-webpack-plugin": "^3.2.0",
-    "tslib": "^2.0.0",
-    "twemoji": "^13.0.1",
-    "typescript": "^4.4.2",
-    "webpack": "^4.41.2",
-    "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.9.0",
-    "webpack-livereload-plugin": "^2.1.1"
+    "@types/jest": "^29.5.11",
+    "@types/node": "^20.10.4",
+    "@types/tinymce": "^4.6.9",
+    "@typescript-eslint/eslint-plugin": "^6.13.2",
+    "@typescript-eslint/parser": "^6.13.2",
+    "@vercel/ncc": "^0.38.1",
+    "eslint": "^8.55.0",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-node": "^11.1.0",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.1",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3",
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4"
   }
 }
