+++ src/main.ts
@@ -1,0 +1,5 @@
+import Plugin from './plugin/main';
+
+Plugin();
+
+/** DO NOT EXPORT ANYTHING - IF YOU DO ROLLUP WILL LEAVE A GLOBAL ON THE PAGE */
