+++ src/plugin/main.ts
@@ -1,0 +1,25 @@
+import { isWordContent, preProcess } from "./WordFilter";
+import type Editor from 'tinymce/core/api/Editor';
+
+interface PreProcessEvent {
+    content: string;
+    internal: boolean;
+}
+  
+export default (): void => {
+    tinymce.PluginManager.add("paste_from_word", (editor: Editor) => {
+        editor.on("PastePreProcess", ({ content }: PreProcessEvent) => {
+            if (isWordContent(content)) {
+                event.content = preProcess(editor, content);
+            }
+        });
+
+        return {
+            getMetadata: () => ({
+                name: "Paste from Word",
+                url: "https://github.com/pangaeatech/tinymce-paste-from-word-plugin"
+            })
+        };
+    });
+}
+
